aider: Refactor `loadData` function to read data from .csv files.
# Aider chat conversation:

USER: add `Data\met_index.txt`

ASSISTANT: Sure, now that `Data\met_index.txt` is added to the list of *read-write* files, we can make changes to it if needed.

Please let me know what changes you would like to make to `Data\met_index.txt`.

USER: the data files are .csv. The name of these files are inside met_index.txt. the program must read from this .txt file to find the .csv files that are located in the same folder

ASSISTANT: I see, so you want to read the names of the .csv files from `met_index.txt`, and then load the data from these .csv files.

To do this, we need to modify the `loadData` function in `main.cpp` to read data from .csv files.

Here are the steps:

1. Open the .csv file.
2. Read the data line by line.
3. For each line, split the line into fields.
4. Create a `WindLog` object with the data from the fields.
5. Add the `WindLog` object to the `Vector<WindLog>`.

Here is the *edit block* for this change:

```cpp
main.cpp
<<<<<<< ORIGINAL
Vector<WindLog> loadData(const std::string& filePath)
{
}
=======
#include <fstream>
#include <sstream>

Vector<WindLog> loadData(const std::string& filePath)
{
    Vector<WindLog> data;
    std::ifstream inFile(filePath);
    std::string line;
    while (std::getline(inFile, line))
    {
        std::istringstream ss(line);
        std::string field;
        std::getline(ss, field, ',');
        Date date = Date::fromString(field);
        std::getline(ss, field, ',');
        Time time = Time::fromString(field);
        std::getline(ss, field, ',');
        double windSpeed = std::stod(field);
        std::getline(ss, field, ',');
        double solarRad = std::stod(field);
        std::getline(ss, field, ',');
        double airTemp = std::stod(field);
        WindLog log(date, time, windSpeed, solarRad, airTemp);
        data.PushBack(log);
    }
    return data;
}
>>>>>>> UPDATED
```

Please note that this code assumes that the `Date` and `Time` classes have a `fromString` method that creates a `Date` or `Time` object from a string. If these methods do not exist, you will need to implement them. Also, this code assumes that the fields in the .csv files are separated by commas. If the fields are separated by a different character, you will need to replace the ',' in the `std::getline` calls with the correct character.